feat: deploy SupaCore schema and structure
