+++ src/backend/test.py
@@ -1,0 +1,1 @@
+print("[USER CHECK]", agent.select("users").data)
