+++ src/frontend/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="refresh" content="0; url=login.html" />
+  <title>Redirecting to Login</title>
+</head>
+<body>
+  <p>Redirecting to login page...</p>
+</body>
+</html>
