--- test.py
@@ -1,1 +1,0 @@
-print("[USER CHECK]", agent.select("users").data)
